+++ src/app/portlets/dot-edit-page/shared/models/dot-layout-column.model.ts
