+++ src/app/portlets/dot-edit-page/shared/models/dot-layout-grid-box.model.ts
@@ -1,5 +1,5 @@
 import { NgGridItemConfig } from 'angular2-grid';
-import { Container } from './container.model';
+import { Container } from './dot-container.model';
 
 export interface LayoutGridBox {
     config: NgGridItemConfig;
