+++ src/app/portlets/dot-edit-page/shared/models/dot-layout-row.model.ts
@@ -1,4 +1,4 @@
-import { LayoutColumn } from './layout-column.model';
+import { LayoutColumn } from './dot-layout-column.model';
 
 export interface LayoutRow {
     columns: LayoutColumn[];
