+++ src/app/portlets/dot-edit-page/shared/models/dot-layout-sidebar.model.ts
@@ -1,0 +1,3 @@
+export interface DotLayoutSideBar {
+
+}
