+++ src/app/portlets/dot-edit-page/shared/models/dot-layout.model.ts
