+++ src/app/portlets/dot-edit-page/shared/models/dot-page-view.model.ts
@@ -1,4 +1,4 @@
-import { Layout} from './layout.model';
+import { Layout} from './dot-layout.model';
 import { Page } from './page.model';
 
 export interface PageView {
