--- src/app/portlets/dot-edit-page/shared/models/layout-sidebar.model.ts
@@ -1,3 +1,0 @@
-export interface LayoutSideBar {
-
-}
